pyproject.toml
[tool.poetry]
name = "area-estimator"
version = "0.1.0"
description = ""
authors = ["Ivan Malison <IvanMalison@gmail.com>"]
packages = [{include = "area_estimator"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
coloredlogs = "^15.0.1"
dependency-injector = "^4.41.0"
ipdb = "^0.13.13"
labelbox = {extras = ["data"], version = "^3.49.1"}
numpy = "^1.24.2"
opencv-python = "^4.7.0.72"
pillow = "^10.0.0"
python-dotenv = "^1.0.0"
shapely = "^2.0.1"
xdg-base-dirs = "^6.0.0"
pyyaml = "^6.0.1"
scipy = "^1.11.1"

[tool.poetry.group.dev.dependencies]
python-lsp-server = {extras = ["all"], version = "^1.7.1"}
pytest = "*"
mock = "*"
flameprof = "*"
mypy = "^1.4.1"
flake8 = "^6.0.0"

[tool.poetry.scripts]
area-estimator = 'area_estimator.console:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = '-p no:warnings'
filterwarning = [
   "ignore::Deprecated",
]
